update: ads controller
