+++ routes/web.php
@@ -1,11 +1,14 @@
 <?php
 
+use App\Http\Controllers\admin\AdController;
 use App\Http\Controllers\advertisement\advertisementController;
 use App\Http\Controllers\advertisement\AdvertisementController as AdvertisementAdvertisementController;
 use App\Http\Controllers\pharmacy\PharmacyController;
 use Illuminate\Support\Facades\Route;
 
 use Barryvdh\Debugbar\Facades\Debugbar;
+use Illuminate\Routing\RouteGroup;
+
 // TODO
 // disable Debug
 Debugbar::disable();
@@ -33,9 +36,14 @@ Route::put('/pharmacies/{id}', [PharmacyController::class, 'update']);
 Route::delete('/pharmacies/{id}', [PharmacyController::class, 'destroy']);
 Route::get('/pharmacies', [PharmacyController::class, 'index']);
 
-Route::get('/advertisement', [AdvertisementController::class, 'index']);
-Route::get('/advertisement/{id}', [AdvertisementController::class, 'show']);
-Route::post('/advertisement', [AdvertisementController::class, 'store']);
-Route::put('/advertisement/{id}', [AdvertisementController::class, 'update']);
-Route::delete('/advertisement/{id}', [AdvertisementController::class, 'destroy']);
-Route::get('/advertisement', [AdvertisementController::class, 'index']);
+
+
+Route::prefix('admin')->group(function () {
+
+  // TODO
+  // Route::middleware(['auth'])->group(function () {
+  /* ads */
+  Route::resource('/ads', AdController::class);
+
+  // });
+});
